requirements.txt
streamlit>=1.25
langchain>=0.0.240
chromadb>=0.4.5
pymupdf>=1.22.5
weasyprint>=58.0
openai>=0.27.8
transformers>=4.30.0
sentence-transformers>=2.2.2
huggingface-hub>=0.16.4
openai-whisper>=1.0
python-docx>=0.8.11
faiss-cpu>=1.7.4
langchain_openai>=0.0.1
langchain_community>=0.0.1
llama-index>=0.8.0
aiohttp>=3.8.4
supabase>=0.0.3
python-dotenv>=1.0.0
pytest>=7.3.1
